+++ src/main/java/refinedstorage/tile/autocrafting/TileCraftingMonitor.java
@@ -9,6 +9,8 @@ import refinedstorage.tile.TileMachine;
 import refinedstorage.tile.autocrafting.task.ICraftingTask;
 
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 
 public class TileCraftingMonitor extends TileMachine {
@@ -73,6 +75,8 @@ public class TileCraftingMonitor extends TileMachine {
                 newTasks.add(task);
             }
         }
+
+        Collections.reverse(newTasks);
 
         tasks = newTasks;
     }