+++ src/main/java/refinedstorage/tile/grid/TileGrid.java
@@ -65,6 +65,8 @@ public class TileGrid extends TileMachine implements IGrid {
     private int sortingType = SORTING_TYPE_NAME;
     private int searchBoxMode = SEARCH_BOX_MODE_NORMAL;
 
+    private EnumGridType type;
+
     private List<ItemGroup> itemGroups = new ArrayList<ItemGroup>();
 
     @Override
@@ -77,11 +79,11 @@ public class TileGrid extends TileMachine implements IGrid {
     }
 
     public EnumGridType getType() {
-        if (worldObj.getBlockState(pos).getBlock() == RefinedStorageBlocks.GRID) {
-            return (EnumGridType) worldObj.getBlockState(pos).getValue(BlockGrid.TYPE);
+        if (type == null && worldObj.getBlockState(pos).getBlock() == RefinedStorageBlocks.GRID) {
+            this.type = (EnumGridType) worldObj.getBlockState(pos).getValue(BlockGrid.TYPE);
         }
 
-        return EnumGridType.NORMAL;
+        return type == null ? EnumGridType.NORMAL : type;
     }
 
     @Override